pyproject.toml
[tool.poetry]
name = "uwu"
version = "0.1"
license = "MIT"
readme = "README.md"
authors = ["LinNian<2544704967@qq.com>"]
description = "uwu"

[tool.poetry.dependencies]
python = "^3.9"
aiocache = ">=0.11.0"
aiofiles = "^0.8.0"
emoji = "^1.7.0"
fonttools = "^4.0.0"
httpx = ">=0.19.0"
imageio = "^2.12.0"
numpy = "^1.20.0"
opencv-python-headless = "^4.0.0"
Pillow = ">=8.2.0"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencent-cloud"
default = true
url = "https://mirrors.cloud.tencent.com/pypi/simple"
